Generic schema and migrations, separate postgres
