+++ crates/nostr-sqldb/src/postgres.rs
@@ -1,6 +1,4 @@
 use deadpool::managed::{Object, Pool};
-use diesel::dsl::{Eq, Filter as DieselFilter, InnerJoin, IntoBoxed};
-use diesel::pg::Pg;
 use diesel::prelude::*;
 use diesel::result::{DatabaseErrorKind, Error as DieselError};
 use diesel::QueryResult;
@@ -9,10 +7,13 @@ use diesel_async::scoped_futures::ScopedFutureExt;
 use diesel_async::{AsyncConnection, AsyncPgConnection, RunQueryDsl};
 use nostr::event::*;
 use nostr::filter::Filter;
+use nostr::types::Timestamp;
 use nostr_database::*;
+use prelude::BoxedFuture;
 
 use super::model::{EventDataDb, EventDb};
 use super::schema::postgres::{event_tags, events};
+use crate::query::{build_filter_query, event_by_id, with_limit};
 
 /// Shorthand for a database connection pool type
 pub type PostgresConnectionPool = Pool<AsyncDieselConnectionManager<AsyncPgConnection>>;
@@ -76,15 +77,122 @@ impl NostrPostgres {
         &self,
         event_id: &EventId,
     ) -> Result<Option<EventDb>, DatabaseError> {
-        let event_id = event_id.to_hex();
-        let res = events::table
-            .select(EventDb::as_select())
-            .filter(events::id.eq(event_id))
+        let res = event_by_id(event_id)
             .first(&mut self.get_connection().await?)
             .await
             .optional()
             .map_err(DatabaseError::backend)?;
         Ok(res)
+    }
+}
+
+impl NostrEventsDatabase for NostrPostgres {
+    /// Save [`Event`] into store
+    ///
+    /// **This method assumes that [`Event`] was already verified**
+    fn save_event<'a>(
+        &'a self,
+        event: &'a Event,
+    ) -> BoxedFuture<'a, Result<SaveEventStatus, DatabaseError>> {
+        Box::pin(async move { self.save(EventDataDb::try_from(event)?).await })
+    }
+
+    /// Check event status by ID
+    ///
+    /// Check if the event is saved, deleted or not existent.
+    fn check_id<'a>(
+        &'a self,
+        event_id: &'a EventId,
+    ) -> BoxedFuture<'a, Result<DatabaseEventStatus, DatabaseError>> {
+        Box::pin(async move {
+            let status = match self.event_by_id(event_id).await? {
+                Some(e) if e.deleted => DatabaseEventStatus::Deleted,
+                Some(_) => DatabaseEventStatus::Saved,
+                None => DatabaseEventStatus::NotExistent,
+            };
+            Ok(status)
+        })
+    }
+
+    /// Coordinate feature is not supported yet
+    fn has_coordinate_been_deleted<'a>(
+        &'a self,
+        _coordinate: &'a nostr::nips::nip01::CoordinateBorrow<'a>,
+        _timestamp: &'a Timestamp,
+    ) -> BoxedFuture<'a, Result<bool, DatabaseError>> {
+        Box::pin(async move { Ok(false) })
+    }
+
+    /// Get [`Event`] by [`EventId`]
+    fn event_by_id<'a>(
+        &'a self,
+        event_id: &'a EventId,
+    ) -> BoxedFuture<'a, Result<Option<Event>, DatabaseError>> {
+        Box::pin(async move {
+            let event = match self.event_by_id(event_id).await? {
+                Some(e) if !e.deleted => {
+                    Some(Event::decode(&e.payload).map_err(DatabaseError::backend)?)
+                }
+                _ => None,
+            };
+            Ok(event)
+        })
+    }
+
+    /// Count the number of events found with [`Filter`].
+    ///
+    /// Use `Filter::new()` or `Filter::default()` to count all events.
+    fn count(&self, filter: Filter) -> BoxedFuture<Result<usize, DatabaseError>> {
+        Box::pin(async move {
+            let res: i64 = build_filter_query(filter)
+                .count()
+                .get_result(&mut self.get_connection().await?)
+                .await
+                .map_err(DatabaseError::backend)?;
+            Ok(res as usize)
+        })
+    }
+
+    /// Query stored events.
+    fn query(&self, filter: Filter) -> BoxedFuture<Result<Events, DatabaseError>> {
+        let filter = with_limit(filter, 10000);
+        Box::pin(async move {
+            let mut events = Events::new(&filter);
+            let result = build_filter_query(filter.clone())
+                .select(EventDb::as_select())
+                .load(&mut self.get_connection().await?)
+                .await
+                .map_err(DatabaseError::backend)?;
+
+            for item in result.into_iter() {
+                if let Ok(event) = Event::decode(&item.payload) {
+                    events.insert(event);
+                }
+            }
+            Ok(events)
+        })
+    }
+
+    /// Delete all events that match the [Filter]
+    fn delete(&self, filter: Filter) -> BoxedFuture<Result<(), DatabaseError>> {
+        let filter = with_limit(filter, 999);
+        Box::pin(async move {
+            let filter = build_filter_query(filter);
+            diesel::update(events::table)
+                .set(events::deleted.eq(true))
+                .filter(events::id.eq_any(filter.select(events::id)))
+                .execute(&mut self.get_connection().await?)
+                .await
+                .map_err(DatabaseError::backend)?;
+
+            Ok(())
+        })
+    }
+}
+
+impl NostrDatabase for NostrPostgres {
+    fn backend(&self) -> Backend {
+        Backend::Custom("Postgres".to_string())
     }
 }
 
@@ -117,82 +225,10 @@ impl std::fmt::Debug for NostrPostgres {
     }
 }
 
-/// sets the given default limit on a Nostr filter if not set
-pub fn with_limit(filter: Filter, default_limit: usize) -> Filter {
-    if filter.limit.is_none() {
-        return filter.limit(default_limit);
+/// For now we want to avoid wiping the database
+impl NostrDatabaseWipe for NostrPostgres {
+    #[inline]
+    fn wipe(&self) -> BoxedFuture<Result<(), DatabaseError>> {
+        Box::pin(async move { Err(DatabaseError::NotSupported) })
     }
-    filter
-}
-
-// filter type of a join query.
-type QuerySetJoinType<'a> = IntoBoxed<
-    'a,
-    DieselFilter<
-        InnerJoin<events::table, event_tags::table>,
-        Eq<event_tags::event_id, diesel::expression::SqlLiteral<diesel::sql_types::VarChar>>,
-    >,
-    Pg,
->;
-
-pub fn build_filter_query<'a>(filter: Filter) -> QuerySetJoinType<'a> {
-    let mut query = events::table
-        .distinct()
-        .inner_join(event_tags::table)
-        .filter(events::deleted.eq(false))
-        .into_boxed();
-
-    if let Some(limit) = filter.limit {
-        query = query.limit(limit as i64);
-    }
-
-    if !has_filters(&filter) {
-        return query;
-    }
-    if let Some(ids) = filter.ids.clone() {
-        let values = ids.iter().map(|id| id.to_hex()).collect::<Vec<_>>();
-        query = query.filter(events::id.eq_any(values));
-    }
-
-    if let Some(authors) = filter.authors.clone() {
-        let values = authors.iter().map(|a| a.to_hex()).collect::<Vec<_>>();
-        query = query.filter(events::pubkey.eq_any(values));
-    }
-
-    if let Some(kinds) = filter.kinds.clone() {
-        let values = kinds.iter().map(|k| k.as_u16() as i64).collect::<Vec<_>>();
-        query = query.filter(events::kind.eq_any(values));
-    }
-
-    if let Some(since) = filter.since {
-        query = query.filter(events::created_at.ge(since.as_u64() as i64));
-    }
-
-    if let Some(until) = filter.until {
-        query = query.filter(events::created_at.le(until.as_u64() as i64));
-    }
-
-    if !filter.generic_tags.is_empty() {
-        for (tag, values) in filter.generic_tags.into_iter() {
-            let values = values.iter().map(|v| v.to_string()).collect::<Vec<_>>();
-            query = query.filter(
-                event_tags::tag
-                    .eq(tag.to_string())
-                    .and(event_tags::tag_value.eq_any(values)),
-            );
-        }
-    }
-
-    query
-}
-
-// determine if the filter has any filters set
-fn has_filters(filter: &Filter) -> bool {
-    filter.ids.is_some()
-        || filter.authors.is_some()
-        || filter.kinds.is_some()
-        || filter.since.is_some()
-        || filter.until.is_some()
-        || !filter.generic_tags.is_empty()
-        || filter.limit.is_some()
 }
